Let compiler show more warnings

+++ src/gwmodelpp/CGwmVariableForwardSelector.cpp
@@ -8,15 +8,15 @@ using namespace std;
 vector<size_t> CGwmVariableForwardSelector::optimize(IGwmVarialbeSelectable *instance)
 {
     vector<size_t> curIndex, restIndex;
-    for (int i = 0; i < mVariables.size(); i++)
+    for (size_t i = 0; i < mVariables.size(); i++)
     {
         restIndex.push_back(i);
     }
     vector<pair<vector<size_t>, double> > modelCriterions;
-    for (int i = 0; i < mVariables.size(); i++)
+    for (size_t i = 0; i < mVariables.size(); i++)
     {
         vec criterions = vec(mVariables.size() - i);
-        for (int j = 0; j < restIndex.size(); j++)
+        for (size_t j = 0; j < restIndex.size(); j++)
         {
             curIndex.push_back(restIndex[j]);
             double aic = instance->getCriterion(convertIndexToVariables(curIndex));
@@ -35,7 +35,7 @@ vector<size_t> CGwmVariableForwardSelector::optimize(IGwmVarialbeSelectable *ins
 std::vector<std::size_t> CGwmVariableForwardSelector::convertIndexToVariables(std::vector<std::size_t> index)
 {
     vector<size_t> variables;
-    for (int i : index)
+    for (size_t i : index)
         variables.push_back(mVariables[i]);
     return variables;
 }
@@ -72,7 +72,7 @@ VariablesCriterionList CGwmVariableForwardSelector::indepVarsCriterion() const
     for (pair<vector<size_t>, double> item : mVarsCriterion)
     {
         vector<size_t> variables;
-        for (int i : item.first)
+        for (size_t i : item.first)
         {
             variables.push_back(mVariables[i]);
         }
